DONE registration with the IMPL But error in pstm query

+++ WebApp1/src/main/java/in/adit/model/Employee.java
@@ -49,7 +49,7 @@ public class Employee {
 		this.email = email;
 	}
 	
-	public int getNumber() {
+	public Long getNumber() {
 		return number;
 	}
